feat: add team test cases

+++ govtech-be/src/services/teamService.ts
@@ -1,7 +1,7 @@
 import { QueryDocumentSnapshot } from "firebase-admin/firestore";
+import { Group } from "src/models/groupModel";
 import { db } from "../configs/firebase";
 import { Team } from "../models/teamModel";
-
 export class TeamService {
   public async get(id: string): Promise<Team | null> {
     const teamSnapshot = await db.collection("teams").doc(id).get();
@@ -9,15 +9,15 @@ export class TeamService {
       return null;
     }
     const teamData = teamSnapshot.data();
-    if (!teamData) {
-      return null;
+    let team: Team | null = null;
+    if (teamData) {
+      team = {
+        id: teamData.id,
+        name: teamData.name,
+        regDate: teamData.regDate,
+        group: teamData.group,
+      };
     }
-    const team: Team = {
-      id: teamData.id,
-      name: teamData.name,
-      regDate: teamData.regDate,
-      group: teamData.group,
-    };
     return team;
   }
 
@@ -28,11 +28,27 @@ export class TeamService {
 
     if (team.id) {
       const teamSnapshot = await teamRef.get();
-      if (!teamSnapshot.exists) {
+      if (!teamSnapshot || (teamSnapshot && !teamSnapshot.exists)) {
         throw new Error(
           `Team with ID ${team.id} does not exist. Please provide a valid ID to update.`
         );
       }
+    }
+
+    const groupRef = db.collection("groups").doc(team.group);
+    const groupSnapshot = await groupRef.get();
+
+    if (groupSnapshot && groupSnapshot.exists) {
+      const groupData = groupSnapshot.data();
+      let currentCount: number = 0;
+      if (groupData) {
+        currentCount = groupData.count;
+      }
+      const newCount = currentCount + 1;
+      await groupRef.update({ count: newCount });
+    } else {
+      const newGroupData: Group = { count: 1 };
+      await groupRef.set(newGroupData);
     }
 
     await teamRef.set({
@@ -63,6 +79,38 @@ export class TeamService {
 
   public async delete(id: string): Promise<void> {
     const teamRef = db.collection("teams").doc(id);
+    const teamSnapshot = await teamRef.get();
+
+    if (!teamSnapshot.exists) {
+      throw new Error(`Team with ID ${id} does not exist.`);
+    }
+
+    const teamData = teamSnapshot.data();
+    let groupId = null;
+    if (teamData) {
+      groupId = teamData.group;
+    }
+
     await teamRef.delete();
+
+    if (groupId) {
+      const groupRef = db.collection("groups").doc(groupId);
+      const groupSnapshot = await groupRef.get();
+
+      if (groupSnapshot.exists) {
+        const groupData = groupSnapshot.data();
+        let currentCount: number = 0;
+        if (groupData) {
+          currentCount = groupData.count;
+        }
+        const newCount = currentCount > 0 ? currentCount - 1 : 0;
+
+        await groupRef.update({ count: newCount });
+
+        if (newCount === 0) {
+          await groupRef.delete();
+        }
+      }
+    }
   }
 }
